login logout register

+++ app/Http/Controllers/api/v1/AuthController.php
@@ -29,12 +29,21 @@ class AuthController extends Controller
      */
     public function index()
     {
-        $users = User::all();
+        $user = Auth::user();
 
-        return response()->json([
-            'status'    => true,
-            'users'     => $users
-        ]);
+        if($user->hasDirectPermission('get user'))
+        {
+            $users = User::all();
+            return response()->json([
+                'status'    => true,
+                'users'     => $users
+            ]);
+        } else {
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User doesnt have permission to this action'
+            ], 200);
+        }
     }
 
     /**
@@ -45,14 +54,27 @@ class AuthController extends Controller
      */
     public function store(Request $request)
     {
+        $user = Auth::user();
 
-        $user = User::create($request->all());
+        if($user->hasDirectPermission('create user'))
+        {
+            $user = $this->user::create([
+                'name' => $request['name'],
+                'email' => $request['email'],
+                'password' => bcrypt($request['password']),
+            ]);
 
-        return response()->json([
-            'status'    => true,
-            'message'   => 'User created successfully',
-            'user'  => $user
-        ], 200);
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User created successfully',
+                'user'  => $user
+            ], 200);
+        } else {
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User doesnt have permission to this action'
+            ], 200);
+        }
     }
 
     /**
@@ -75,13 +97,23 @@ class AuthController extends Controller
      */
     public function update(Request $request, User $user)
     {
-        $user->update($request->all());
+        $user = Auth::user();
 
-        return response()->json([
-            'status'    => true,
-            'message'   => 'User updated successfully',
-            'user'  => $user
-        ], 200);
+        if($user->hasDirectPermission('edit user'))
+        {
+            $user->update($request->all());
+
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User updated successfully',
+                'user'  => $user
+            ], 200);
+        } else {
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User doesnt have permission to this action'
+            ], 200);
+        }
     }
 
     /**
@@ -93,29 +125,75 @@ class AuthController extends Controller
     public function destroy(User $user)
     {
 
-        if(Auth::user()->id == $user->id){
+        $user = Auth::user();
+
+        if($user->hasDirectPermission('delete user'))
+        {
+            if(Auth::user()->id == $user->id){
+                return response()->json([
+                    'status'    => true,
+                    'message'   => 'Cannot delete your own user'
+                ], 401);
+            }
+
+            $user->delete();
             return response()->json([
                 'status'    => true,
-                'message'   => 'Cannot delete your own user'
-            ], 401);
+                'message'   => 'User deleted successfully'
+            ], 200);
+        }else {
+            return response()->json([
+                'status'    => true,
+                'message'   => 'User doesnt have permission to this action'
+            ], 200);
         }
 
-        $user->delete();
-        return response()->json([
-            'status'    => true,
-            'message'   => 'User deleted successfully'
-        ], 200);
-
 
     }
 
-
+    // Register new user
     public function register(Request $request)
     {
-        return 1;
+        // validate the incoming request
+        // set every field as required
+        // set email field so it only accept the valid email format
+
+        $this->validate($request, [
+            'name' => 'required|string|min:2|max:255',
+            'email' => 'required|string|email:rfc,dns|max:255|unique:users',
+            'password' => 'required|string|min:6|max:255',
+        ]);
+
+        // if the request valid, create user
+
+        $user = $this->user::create([
+            'name' => $request['name'],
+            'email' => $request['email'],
+            'password' => bcrypt($request['password']),
+        ]);
+
+        // login the user immediately and generate the token
+        $token = auth()->login($user);
+
+        // return the response as json
+        return response()->json([
+            'meta' => [
+                'code' => 200,
+                'status' => 'success',
+                'message' => 'User created successfully!',
+            ],
+            'data' => [
+                'user' => $user,
+                'access_token' => [
+                    'token' => $token,
+                    'type' => 'Bearer',
+                    'expires_in' => auth()->factory()->getTTL() * 60,    // get token expires in seconds
+                ],
+            ],
+        ]);
     }
 
-    // Validation method
+    // JWT Validation method
     public function login(Request $request)
     {
         $this->validate($request, [
@@ -150,6 +228,27 @@ class AuthController extends Controller
             ]);
         }else {
             return response()->json(['status' => false, 'message' => 'Invalid credentials'], 401);
+        }
+    }
+
+    // Logout user
+    public function logout()
+    {
+        // get token
+        $token = JWTAuth::getToken();
+
+        // invalidate token
+        $invalidate = JWTAuth::invalidate($token);
+
+        if($invalidate) {
+            return response()->json([
+                'meta' => [
+                    'code' => 200,
+                    'status' => 'success',
+                    'message' => 'Successfully logged out',
+                ],
+                'data' => [],
+            ]);
         }
     }
